change:Team CRUD Complate && about page show all team with header && footer section website Setting show

+++ app/Http/Controllers/TeamController.php
@@ -39,26 +39,39 @@ class TeamController extends Controller
      */
     public function store(Request $request)
     {
-         //User Validation
+         //Team Validation
       $this->validate($request,[
-         'name' => 'required|string|max:255',
-         'email' => "required|email|unique:users,email",
-         'password' => 'required|min:8',
-         'image' => 'sometimes|nullable|image|max:2048',
+         'team_name' => 'required|string|unique:teams,team_name',
+         'designation' => 'required|string',
+         'team_description' => "required",
+         'team_picture' => 'sometimes|nullable|image|max:2048',
       ]);
-      //user insert
+      //Team insert
       $team = Team::create([
-         'team_name' => $request->name,
-         'team_img' => $request->email,
-         'team_desc' => bcrypt($request->password),
-         'team_facebook_link' => $request->desc,
-         'team_twitter_link' => Str::slug($request->name),
-         'team_linkdin_link' => 0,
+         'team_name' => $request->team_name,
+         'designation' => $request->designation,
+         'team_desc' => $request->team_description,
+         'team_img' =>  'backend/blog/default.png',
+         'team_facebook_link' => $request->team_facebook,
+         'team_twitter_link' => $request->team_twitter,
+         'team_linkdin_link' => $request->team_linkdin,
 
       ]);
+
+      if($request->has('team_picture')) {
+
+      //    $imageName = time().'.'.$request->post_picture->extension();  
+      //    $request->post_picture->move(public_path('backend/blog/'), $imageName);
+      //   return $post->imageName;
+
+         $filename = time() . '.' .$request->team_picture->getClientOriginalextension();
+         //$request->post_picture->move('backend/blog/', $filename);
+         $request->team_picture->move(public_path('backend/team/'), $filename);
+         $team->team_img = 'backend/team/'.$filename;
+         // $team->save();
+      }
 
       $team->save();
-
       Session::flash('status','Team Created Successfully');
       return redirect()->route('team.index');
     }
@@ -96,7 +109,41 @@ class TeamController extends Controller
      */
     public function update(Request $request, Team $team)
     {
-     // return view('backend.team.edit');
+     
+    //  return $request->all();
+      //Team Validation
+      $this->validate($request,[
+         'team_name' => 'sometimes|nullable|required|string',
+         'designation' => 'sometimes|nullable|required|string',
+         'team_description' => "required",
+         'team_picture' => 'sometimes|nullable|image|max:2048',
+      ]);
+
+      //Team Update
+      $team->team_name = $request->team_name; 
+      $team->designation = $request->designation; 
+      $team->team_desc = $request->team_description; 
+      $team->team_facebook_link = $request->team_facebook; 
+      $team->team_twitter_link = $request->team_twitter; 
+      $team->team_linkdin_link = $request->team_linkdin; 
+
+       //Team Picture Change
+         if($request->has('team_picture')) {
+
+         //    $imageName = time().'.'.$request->post_picture->extension();  
+         //    $request->post_picture->move(public_path('backend/blog/'), $imageName);
+         //   return $post->imageName;
+
+            $filename = time() . '.' .$request->team_picture->getClientOriginalextension();
+            //$request->post_picture->move('backend/blog/', $filename);
+            $request->team_picture->move(public_path('backend/team/'), $filename);
+            $team->team_img = 'backend/team/'.$filename;
+            // $team->save();
+         }
+
+      $team->save();
+      Session::flash('status','Team Information Changed Successfully');
+      return redirect()->route('team.index');
 
     }
 
@@ -108,7 +155,15 @@ class TeamController extends Controller
      */
     public function destroy(Team $team)
     {
-        $team->delete();
-        return 'Team Delete';
+       
+        //Team Delete
+        if($team){
+         $team->delete();
+        }
+        else{
+         abort(404);
+        }
+        Session::flash('error','Team Delete Successfully');
+        return redirect()->route('team.index');
     }
 }
